+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- [fj-doc-maven-plugin] updated readme information for quarkus-3 flavour
 - [fj-doc-guide] added pdf documentation
 - [fj-doc-playground-quarkus] fj-doc-ext-kotlin-version set to 0.4.2